Moving from unsigned short to uint32_t.

+++ search/minibs/minibs.hpp
@@ -50,7 +50,7 @@ public:
     // This sparsification is not easily done on the fly, so we do it during
     // postprocessing.
 
-    flat_hash_map<uint32_t, unsigned short> fingerprint_map;
+    flat_hash_map<uint32_t, uint32_t> fingerprint_map;
     // We store pointers to the winning fingerprints,
     // to enable smooth sparsification.
     std::vector<flat_hash_set<unsigned int> *> fingerprints;
